requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
distro==1.9.0
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.7
jiter==0.10.0
nodeenv==1.9.1
ollama==0.2.1
openai==1.98.0
pip==24.0
platformdirs==4.2.2
pre-commit==3.7.1
pydantic==2.7.4
pydantic-core==2.18.4
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.4
types-requests==2.32.0.20240602
typing-extensions==4.12.2
typing-inspection==0.4.1
urllib3==2.2.1
uv==0.2.5
virtualenv==20.26.2
